inventory-agents/src/inventory_agents/model.py
import os
from dotenv import load_dotenv
from openai import AsyncOpenAI
from agents import OpenAIChatCompletionsModel

# Load .env file
load_dotenv()

# Read values
GEMINI_MODEL = os.getenv("GEMINI_MODEL")
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
GEMINI_BASE_URL = os.getenv("GEMINI_BASE_URL")

# Setup Gemini client
client = AsyncOpenAI(api_key=GEMINI_API_KEY, base_url=GEMINI_BASE_URL)

# Create model
model = OpenAIChatCompletionsModel(model=GEMINI_MODEL, openai_client=client)
