inventory-agents/pyproject.toml
[project]
name = "inventory-agents"
version = "0.1.0"
description = "Inventory management with OpenAI agents"
readme = "README.md"
authors = [
    { name = "NEHA SHAHZAD", email = "shahzadnaha@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.8",
]

[project.scripts]

start-inventory = "inventory_agents.main:start"



[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
inventory-agents = { path = "src" }
